bhav shell script

+++ src/main/resources/bhavdownload.sh
@@ -21,19 +21,27 @@ dateRangeFunction()
         stringGen $startDate
         curl  -o cm$dateString$suffix "$url$yr/$mnt/cm$dateString$suffix"
         if [ $? -eq 0 ]; then
+        	if [ -f ./cm$dateString$suffix ]; then
                 unzip ./cm$dateString$suffix
                 rm -f ./cm$dateString$suffix
+            fi
         fi
 
         startDate=$(date -d "$startDate + 1 days" +%Y%m%d)
       done
 }
 
+if [ $# -ne 3 ]; then
+	echo "usage => bash -x ./bhavdownload.sh 20170101 20170201 <target-directory-in-unix-style>"
+	exit 1
+fi
+
 export url=https://nse-india.com/content/historical/EQUITIES/
 export suffix=bhav.csv.zip
 export outPutFile=myfile.zip
 dateRangeFunction $1 $2
-cp $PWD/*.csv $3
+mv $PWD/*.csv $3
 tail -q -n +2 $3/cm*.csv > $3/output.csv
+rm $3/cm*.csv
 exit 0
